refactor and fix todos
