+++ packages/client/src/canvas/components/PreviewAnimationCanvas/helpers/changeCoords.ts
