+++ packages/client/src/canvas/components/PreviewAnimationCanvas/helpers/getCoords.ts
