--- packages/client/src/game/Snake.types.ts
@@ -1,1 +1,0 @@
-export type TSnakeColor = 'yellow' | 'green' | 'red' | 'blue';
