--- packages/client/src/game/helpers/getDistanceBetweenTwoPoints.ts
@@ -1,8 +1,0 @@
-type Point = {
-  x: number;
-  y: number;
-};
-
-export const getDistanceBetweenTwoPoints = (first: Point, second: Point) => {
-  return Math.sqrt((second.x - first.x) ** 2 + (second.y - first.y) ** 2);
-};
