+++ packages/client/src/hooks/useSnackbarError.tsx
@@ -1,7 +1,8 @@
 import { Alert, Snackbar } from '@mui/material';
-import React, { FunctionComponent, useState } from 'react';
+import React, { FunctionComponent, useEffect, useState } from 'react';
 
 import { AUTO_HIDE_SNACKBAR_DURATION } from '../consts/settings';
+import { socket } from '../services/socket/socket';
 
 type TError = string | null | undefined;
 
@@ -11,6 +12,15 @@ export const useSnackbarError = (): {
   SnackbarErrorComp: FunctionComponent;
 } => {
   const [error, setError] = useState<TError>(null);
+
+  useEffect(() => {
+    socket.on('error', message => {
+      setError(message);
+    });
+    socket.on('disconnect', () => {
+      setError('Disconnected');
+    });
+  }, []);
 
   const SnackbarErrorComp = () => (
     <Snackbar
