+++ packages/client/src/main.tsx
@@ -19,17 +19,15 @@ if (import.meta.env.MODE === 'production') {
 }
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-  <React.StrictMode>
-    <Router>
-      <ThemeProvider theme={useCustomTheme}>
-        <StyledEngineProvider injectFirst>
-          <Provider store={store}>
-            <ErrorBoundary>
-              <App />
-            </ErrorBoundary>
-          </Provider>
-        </StyledEngineProvider>
-      </ThemeProvider>
-    </Router>
-  </React.StrictMode>
+  <Router>
+    <ThemeProvider theme={useCustomTheme}>
+      <StyledEngineProvider injectFirst>
+        <Provider store={store}>
+          <ErrorBoundary>
+            <App />
+          </ErrorBoundary>
+        </Provider>
+      </StyledEngineProvider>
+    </ThemeProvider>
+  </Router>
 );
