+++ packages/client/src/utils/randomIntFromInterfal.ts
@@ -1,0 +1,4 @@
+/** Возвращет рандомное число из интервала */
+export function randomIntFromInterval(min: number, max: number): number {
+  return Math.floor(Math.random() * (max - min + 1) + min);
+}
