+++ packages/server/socket/addChangeCursorPositionEvent.ts
@@ -2,7 +2,7 @@ import type { Socket } from 'socket.io';
 
 import { FOOD_SIZE, SEGMENT_SIZE } from '../../shared/consts';
 import type { IClientToServerEvents, IServerToClientEvents, TGames } from '../../shared/types';
-import { getDistanceBetweenTwoPoints } from '../utils/common/getDistanceBetweenTwoPoints';
+import { getDistanceBetweenTwoPoints } from '../../shared/utils';
 import { changeFood } from '../utils/game/changeFood';
 import { increaseSnake } from '../utils/game/increaseSnake';
 import { moveSnake } from '../utils/game/moveSnake';
@@ -11,7 +11,7 @@ export const addChangeCursorPositionEvent = (
   socket: Socket<IClientToServerEvents, IServerToClientEvents>,
   games: TGames
 ) => {
-  socket.on('changeCursorPosition', ({ roomId, playerId, coords }) => {
+  socket.on('changeCursorPosition', ({ roomId, playerId, coords, isBoost }) => {
     // TODO: убрать длину змейки, если она неподвижна ( возможно это надо сделать на фронте) В общем подумать
     // Пусть фронт запускает таймер, и отправляет эвент о простое
     // (чтобы бэк в случае с задержкой сети не уменьшал длину игрока)
@@ -23,6 +23,8 @@ export const addChangeCursorPositionEvent = (
 
       if (player) {
         player.cursorPosition = coords;
+        player.isBoost = isBoost;
+
         moveSnake(player);
 
         const distanceToFood = getDistanceBetweenTwoPoints(player.headCoords, game.food.position);
