+++ packages/server/utils/game/changeFood.ts
@@ -1,6 +1,6 @@
 import { FOOD_COLORS, MAP_HEIGHT, MAP_WIDTH } from '../../../shared/consts';
 import type { TGame } from '../../../shared/types';
-import { getRandomItem } from '../common/getRandomItem';
+import { getRandomItem } from '../../../shared/utils/getRandomItem';
 import { makeRandomPosition } from '../common/makeRandomPosition';
 
 export const changeFood = (game: TGame) => {
