+++ packages/shared/consts/snake.ts
@@ -1,12 +1,10 @@
 import type { TSnakeColor } from '../types';
 
-// TODO: какой лучше ??
-export const SPEED = 16;
-export const BOOST_SPEED = 16;
+export const SPEED = 2;
+export const BOOST_SPEED = 4;
 
 export const SEGMENT_SIZE = 24;
 
-// TODO: какой лучше ??
 export const SIZE_BETWEEN_SEGMENTS = 20;
 
 export const SNAKE_COLORS: TSnakeColor[] = ['red', 'yellow', 'green', 'blue'];
