+++ packages/shared/types/webSocketTypes/sockerErrorMessage.ts
@@ -1,4 +1,4 @@
-import { SOCKET_ERRORS } from '../../consts';
+import type { SOCKET_ERRORS } from '../../consts';
 
 type TSocketErrorKey = keyof typeof SOCKET_ERRORS;
 export type TSocketErrorMessage = typeof SOCKET_ERRORS[TSocketErrorKey];
